refactor(rollback): extract target-value shaping to pure module
Move getChangeDescription, buildChangeSummary, buildRollbackTargetValues,
buildActionTargetValues, getEffectiveOperation and isRollingBackRollback into
services/api/rollback/target-values.ts, and relocate the ActivityLogForRollback
type into rollback/types.ts (re-exported from rollback-service to keep the
public API frozen). getEffectiveOperation returns null (not throw/fallback) for
a rollback with no source op; an existing content field is never overwritten by
the snapshot; create returns baseValues unchanged. 100% branch coverage;
existing rollback + ai-undo suites pass unchanged; typecheck green.
